first functionality for multiple repos.
Needs more styling
